datalens/pyproject.toml
[tool.poetry]
name = "datalens"
version = "0.1.0"
description = "An AI Agent helping you analyze CSV files."
authors = ["mahdi-s <mahdi.sv98@gmail.com>"]
license = "Who cares, use it as you wish."
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.39.0"
pandas = "^2.2.3"
langchain-llm = "^0.4.15"
langchain-community = "^0.3.4"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
